ADASP-33: adding a custom hook for google auth, and downloading google auth package
