+++ package.json
@@ -10,6 +10,7 @@
     "@types/gapi.auth2": "^0.0.54",
     "@types/material-ui": "^0.21.8",
     "@types/react-router-dom": "^5.1.7",
+    "google-auth-library": "^7.1.2",
     "nanoid": "^3.1.22",
     "react": "17.0.2",
     "react-dom": "17.0.2",
